added interrupts and implemented blind timers

+++ Server/Blinds.go
@@ -1,15 +1,15 @@
 package Server
 
 type BlindLevel struct {
-	sb   int
-	ante int
+	sb   float64
+	ante float64
 }
 
 type Blinds struct {
 	blinds []BlindLevel
 }
 
-func (this *Blinds) Init(sbs, antes []int) {
+func (this *Blinds) Init(sbs, antes []float64) {
 	levels := len(sbs)
 	this.blinds = make([]BlindLevel, levels)
 	for i := 0; i < levels; i++ {
@@ -18,10 +18,10 @@ func (this *Blinds) Init(sbs, antes []int) {
 	}
 }
 
-func (this Blinds) GetSB(level int) int {
+func (this Blinds) GetSB(level int) float64 {
 	return this.blinds[level].sb
 }
 
-func (this Blinds) GetAnte(level int) int {
+func (this Blinds) GetAnte(level int) float64 {
 	return this.blinds[level].ante
 }
